[NETBEANS-6081] - Cleaning module enterprise part 1: (#3221)
- Remove unnecessary boxing
- Remove unnecessary unboxing
- Remove boxing of already boxed value
- Remove creating new Boolean

+++ enterprise/web.struts/src/org/netbeans/modules/web/struts/editor/StrutsConfigHyperlinkProvider.java
@@ -128,7 +128,7 @@ public class StrutsConfigHyperlinkProvider implements HyperlinkProvider {
     public void performClickAction(javax.swing.text.Document doc, int offset) {
         if (debug) debug(":: performClickAction");
         if (hyperlinkTable.get(eav[0]+"#"+eav[1])!= null){
-            int type = ((Integer)hyperlinkTable.get(eav[0]+"#"+eav[1])).intValue();
+            int type = ((Integer)hyperlinkTable.get(eav[0]+"#"+eav[1]));
             switch (type){
                 case JAVA_CLASS: findJavaClass(eav[2], doc); break;
                 case FORM_NAME: findForm(eav[2], (BaseDocument)doc);break;
